Store DWD MOSMIX Forecasts As ASCII File

Writes a dwdforecast object (see get_forecasts) object into an ASCII file.

write_ascii(x, ...)

# S3 method for dwdforecast
write_ascii(x, file = NULL, dir = "DWDMOS",
  parameter = NULL, ...)

Arguments

x
an object of class dwdforecast as returned by get_forecasts.

...
forwarded to S3 methods (of main interest: write_ascii.dwdforecast).

file
NULL or character (name of the output file). If NULL the output file name will be created automatically based on the meta information of the object x.

dir
character, name of the output directory.

parameter
default is NULL. A vector of parameter names can be provided if only a subset of x should be written into the output file.

Details

Method to store DWD MOSMIX forecasts as ASCII. The data will be written into the dir directory. Please onte that this directory has to be existing. If not, the script will stop.

If a file name is provided this user defined name will be used to store the data. If no file argument is provided the method takes care of the output structure/file names as follows:

Create a folder inside dir for each station.

Create one file per DWD MOSMIX initialization time.

Existing files will be replaced.

The default output file name format is DWDMOS_<YYYYmmddHHMMSS>_<station>.dat.

TODO: I should ignore the dir if file is provided.